[master]: Narrow t2 selection criteria.
This narrows the match for things considered to be t2 instances to only
those types that start with "t2.". Previously there was a class of
(invalid) instance types, such as "2b.something" or "t5.something" that
would cause an error on boolean parsing by matching one regex and not
the other.