Support lifecycle customizations
https://www.terraform.io/docs/configuration/resources.html#lifecycle-lifecycle-customizations

My use case is always using the latest AMI available but not automatically replacing the instance
without manually tainting it.

I was hoping to support the lifecycle with a single variable but it doesn't seem syntactically possible
with HCL2 at this time. Instead I'm using one variable for each of the three possible lifecycle meta-arguments.