Provide a toggle to disable the volume_tags
Since lifecycle is a dynamic property, we have to create a separate
resource that has this feature disabled. It's not the most graceful
workaround, but it works well enough.

The default will keep the current behavior unchanged. Any users who
desire to enable this on existing resources would need to perform a
`state mv` to avoid replacing resources.