fix: Make root_block_device.tags default to null
Fixes #440. Without it, if root_block_device is defined without a tags
key, and if enable_volume_tags isn't set to false, a conflict is created
because main.tf:219 defines volume_tags to be at least { Name = ... }
and the root block device gets tags as {}. Making the default null works
around that.